1-next-26/backend/requirements.txt
asgiref==3.11.0
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==24.1.1
packaging==26.0
pillow==12.1.0
sqlparse==0.5.5
